requirements.txt
# cait-whisper — core dependencies
# Install: pip install -r requirements.txt

# ASR engines
faster-whisper
useful-moonshine-onnx
tokenizers

# Audio
sounddevice
soundfile
numpy

# Keyboard & clipboard
keyboard
pyperclip

# System tray
pystray
Pillow

# LLM cleanup (optional - only used when "LLM Cleanup: ON" is toggled)
# Also used by COMMAND mode for the voice-command classifier fallback
ollama

# UI Automation for selection / context detection (used by COMMAND mode)
pywinauto

# Local OCR for screen-context capture (v2.3).
# Only loaded when "Screen Context: ON" is toggled in COMMAND mode.
# ~50 MB install. Comment this line out if you don't want the feature.
rapidocr-onnxruntime
